[bd-1rj] Fix aggregated period summary query and complete E2E tests
- Fixed bug in period_summary_repo: SUM() of BIGINT columns returns NUMERIC,
  needs explicit ::BIGINT cast in aggregated snapshot query
- Added debug output to currency filter test to diagnose 500 errors
- All 6 HTTP E2E tests now passing:
  ✅ Snapshot-based responses
  ✅ Computed from balances fallback
  ✅ Currency filtering (USD, EUR, GBP, multi-currency aggregate)
  ✅ Error handling (400/404)
  ✅ Performance guard (< 500ms, proves no journal_lines scan)
  ✅ JSON DTO structure validation

Note: Also restored report indexes migration (Track D) to fix container startup

+++ modules/gl/src/repos/period_summary_repo.rs
@@ -135,8 +135,8 @@ async fn find_aggregated_snapshot(
             period_id,
             SUM(journal_count)::INTEGER as journal_count,
             SUM(line_count)::INTEGER as line_count,
-            SUM(total_debits_minor) as total_debits_minor,
-            SUM(total_credits_minor) as total_credits_minor,
+            SUM(total_debits_minor)::BIGINT as total_debits_minor,
+            SUM(total_credits_minor)::BIGINT as total_credits_minor,
             MAX(created_at) as latest_created_at
         FROM period_summary_snapshots
         WHERE tenant_id = $1
